+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [2.1.1] - 2026-03-09
+### Added
+- AllaganLib.GameSheets
+  - Added AsTypedRowRef and AsUntypedRowRef as extension methods.
+
 ## [2.1.0] - 2026-03-09
 ### Added
 - AllaganLib.GameSheets